[hir2mpl]Internal commit msg:     !1044 [hir2mpl] Attributed Statement handling     [hir2mpl] clean up pclint alarm in optimize [mapleall]Internal commit msg:     [mplall] Changes to fix pc-lint warning 1762     !1058 [me][big_endian] Fix bug related to big-endian in merge statements optimization     [mapleall]Static check alarm cleaning(6)     [mplcg]Enable SimplifyMulArithmeticPattern     [mplcg] add cgpostpeephole phase     [mapleall] refactor abi

+++ src/hir2mpl/optimize/include/ror.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) [2021] Huawei Technologies Co.,Ltd.All rights reserved.
+ * Copyright (c) [2021-2022] Huawei Technologies Co.,Ltd.All rights reserved.
  *
  * OpenArkCompiler is licensed under Mulan PSL v2.
  * You can use this software according to the terms and conditions of the Mulan PSL v2.
@@ -36,7 +36,7 @@ class Ror {
   const UniqueFEIRExpr &rExpr;
   UniqueFEIRExpr rShiftBaseExpr;
   UniqueFEIRExpr lShiftBaseExpr;
-  uint32 constVal = 0;
+  uint64 constVal = 0;
   uint32 bitWidth = 0;
 };
 } // namespace maple
